error

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -11,7 +11,7 @@ int print_last_digit(int n)
 
 	if (n < 0)
 		n *= -1;
-	last_digit = n % 10;
+	last_digit = (long)n % 10;
 	_putchar('0' + last_digit);
 	return (last_digit);
 }
